Fix REORGANIZE=true not redistributing randomly-distributed tables
We made an optimization (1dc3e975b8a) in which we skip redistribution
in an INSERT query if the target table is randomly-distributed.
In normal cases that makes sense, but in case of ALTER TABLE SET
(REORGANIZE=true) DISTRIBUTED RANDOMLY which internally might use INSERT
to generate redistributed data, the expectation is that the redistribution
should happen and table data should be randomly distributed again.

So now adding a GUC gp_force_random_redistribution to force the
redistribution if wished (i.e. for REORGANIZE=true or for whatever
other reason). The default is set to false to keep the aforementioned
optimization unaffected.

Note that, other cases of (REORGANIZE=true) which internally use CTAS
instead of INSERT will not be affected by this change because the
aforementioned optimization does not affect CTAS (i.e. random re-
distribution always redistributed in CTAS).

Fix #15733.