Fix gp_stat_bgwriter which shows incomplete/incorrect results
The gp_stat_bgwriter view which is supposed to show cluster-wide view
of pg_stat_bgwriter but only shows part of the segments, e.g.:

```sql
postgres=# select * from gp_stat_bgwriter;
 segment_id | checkpoints_timed | checkpoints_req | checkpoint_write_time | checkpoint_sync_time | buffers_checkpoint | buffers_clean | maxwritten_clean | buffers_backend | buffers_backend_fsync | buffers_alloc |          stats_reset
------------+-------------------+-----------------+-----------------------+----------------------+--------------------+---------------+------------------+-----------------+-----------------------+---------------+-------------------------------
          0 |              1097 |               0 |                 37946 |                   15 |                378 |             0 |                0 |            5133 |                     0 |           387 | 2023-04-28 12:46:24.938582-07
         -1 |              1097 |               0 |                 37946 |                   15 |                378 |             0 |                0 |            5133 |                     0 |           387 | 2023-04-28 12:46:24.938582-07
(2 rows)
```
And even the result of the only primary segment is not correct - it is
just repeating the result from the coordinator.

The root cause is that because pg_stat_bgwriter is selecting from a bunch
of STABLE functions, they are being evaluated and simplied in planning phase.
As a result, the coordinator's results are incorrectly used for the primary.
In addition, the subquery SELECT * FROM pg_stat_bgwriter is given a locus
CdbLocusType_General which results in direct dispatch and only one primary
is being dispatched to.

The initial fix that I thought was to not simplify stable functions at
all, but that would be an overkill. And it's not easy to do something like
in #14499 where we skip the simpifying just for the functions that
pg_stat_bgwriter needs - there are too many functions and no clear pattern to
distinguish them.

So just do a simpler fix which is to have a volatile function to select the
results, and let pg_stat_bgwriter select from that function.

Also modified the regress test to be able to catch such issue in future.
Found a new issue with pg_replication_slot which is unrelated to the current
one. Added a FIXME for it.

Fix #15455