Move  JIT GUCs of ORCA to 'sync_guc_name.h' (#15726)
* Move ORCA JIT GUCs to 'sync_guc_name.h' & other improvements

Background

In the PR #15430, new GUCs were introduced in ORCA,
to enable triggering of JIT compilation based on cost model of ORCA.

These GUCs were placed in the “unsync_guc_name.h” , considering they are not accessed
by the Query Executor (QE), but actually these GUCs are accessed by the QE via provider_init().

Improvements

1) To enable synchronization of JIT GUCs of ORCA between query dispatcher and executor,
they are now placed in “sync_guc_name.h”, as these GUCs are accessed by the QE via provider_init().

2) Removed "Optimizer_Jit" ORCA GUC. Now JIT compilation of ORCA & Planner generated
plan is controlled by a common 'JIT' GUC.

Separate GUCs to enable JIT in ORCA and Planner were leading to some extra
cases which could lead to confusion. For example, Consider the case if JIT is enabled in ORCA,
but not in planner. Now if ORCA fails to produce a plan and falls back to Planner, then we
will have an execution plan but not with JIT as it was not turned ON for Planner.
Thus to avoid such cases, it was decided to keep a single GUC to enable JIT in ORCA and Planner.

3) The pipeline code was updated to include ORCA jit GUC "optimizer_jit_above_cost" for testing.


---------

Co-authored-by: nishant sharma <nishants4@nishants46M6RY.vmware.com>